+++ src/samtools/index/test_data/script.sh
@@ -6,9 +6,7 @@ if [ ! -d /tmp/idxstats_source ]; then
 fi
 
 cp -r /tmp/idxstats_source/bio/samtools/idxstats/test/mapped/* src/samtools/idxstats/test_data
-# samtools idxstats a.sorted.bam > a.sorted.idxstats
+# samtools index a_ref.sorted.bam -o a_ref.sorted.bam.bai
+# samtools index a_ref.sorted.bam -c a_ref.sorted.bam.csi
 
-# dowload test data from nf-core module
-wget https://github.com/nf-core/test-datasets/raw/modules/data/genomics/sarscov2/illumina/bam/test.paired_end.sorted.bam
-wget https://github.com/nf-core/test-datasets/raw/modules/data/genomics/sarscov2/illumina/bam/test.paired_end.sorted.bam.bai
-# samtools idxstats test.paired_end.sorted.bam > test_ref.paired_end.sorted.idxstats
+