+++ CHANGELOG.md
@@ -35,6 +35,9 @@
     - `salmon/salmon_index`: Create a salmon index for the transcriptome to use Salmon in the mapping-based mode (PR #24).
     - `salmon/salmon_quant`: Transcript quantification from RNA-seq data (PR #24).
 
+* `samtools`: 
+    - `samtools/index`: Index SAM/BAM/CRAM files.
+
 ## MAJOR CHANGES
 
 ## MINOR CHANGES